added edit-transaction page

+++ src/shared/utils/firebase/firebase.repository.ts
@@ -1,4 +1,4 @@
-import { getDocs, collection, doc, query, where, addDoc, updateDoc, deleteDoc, CollectionReference, QueryFieldFilterConstraint } from 'firebase/firestore'
+import { getDocs, collection, query, where, addDoc, updateDoc, deleteDoc, CollectionReference, QueryFieldFilterConstraint } from 'firebase/firestore'
 import { Repository } from '../database'
 import { generateId } from '../generate-id'
 import { Entity } from '@/shared/types/entity'
@@ -9,7 +9,7 @@ export class FirebaseRepository<T extends Entity = Entity, DTO = Omit<T, 'id'>> 
   private readonly collection: CollectionReference
 
   constructor(
-    private readonly collectionKey: string,
+    collectionKey: string,
     private readonly entityCreator: (dto: DTO) => Entity = this.getEntityDataByDTO
   ) {
     this.collection = collection(db, collectionKey)
@@ -38,14 +38,18 @@ export class FirebaseRepository<T extends Entity = Entity, DTO = Omit<T, 'id'>> 
   }
 
   async updateById(id: Id, dto: Partial<DTO>): Promise<T | null> {
-    const docRef = doc(db, this.collectionKey, id)
+    const docRef = await this.getDocumentRefByEntityId(id)
+    if (!docRef) return null
+
     await updateDoc(docRef, dto as any)
 
     return this.getById(id)
   }
 
   async deleteById(id: Id): Promise<T | null> {
-    const docRef = doc(db, this.collectionKey, id)
+    const docRef = await this.getDocumentRefByEntityId(id)
+    if (!docRef) return null
+
     const entity = await this.getById(id)
 
     await deleteDoc(docRef)
@@ -61,5 +65,13 @@ export class FirebaseRepository<T extends Entity = Entity, DTO = Omit<T, 'id'>> 
     const filterEntries = filterKeys.map(key => where(key, '==', (filters as any)[key]))
 
     return filterEntries
+  }
+
+  private async getDocumentRefByEntityId(id: Id) {
+    const q = query(this.collection, ...this.mapFiltersToFirebaseEntries({ id } as Partial<T>))
+    const querySnapshot = await getDocs(q)
+
+    const documentReferences = querySnapshot.docs.map(doc => doc.ref)
+    return documentReferences[0] ?? null
   }
 }
